% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuthrieMisoData-package.R
\docType{package}
\name{GuthrieMisoData-package}
\alias{GuthrieMisoData}
\alias{GuthrieMisoData-package}
\title{GuthrieMisoData: Data Package of the Drosophila Midgut expression from Marianes et al. 2013}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Javier Carpinteyro-Ponce \email{jcarpinteyro@carnegiescience.edu} (\href{https://orcid.org/0009-0005-2865-1475}{ORCID})

Authors:
\itemize{
  \item Katie Cox \email{katherinecox@jhu.edu}
  \item Sayumi York \email{joclee@bu.edu}
  \item Fred Tan \email{tan@carnegiescience.edu}
}

}
\keyword{internal}
