% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{miso_counts}
\alias{miso_counts}
\title{Data from Guthrie et al. (2022)}
\format{
A phyloseq object with the following slots:
\describe{
\item{otu_table}{Abundance table (OTUs or ASVs)}
\item{tax_table}{Taxonomic classification}
\item{sam_data}{Sample metadata}
\item{phy_tree}{Phylogenetic tree (if present)}
\item{refseq}{Reference sequences (if present)}
}
}
\source{
\url{https://doi.org/10.1016/j.chom.2022.05.003}
}
\usage{
miso_counts
}
\description{
Microbiome individuality and stability over time - counts
}
\keyword{datasets}
